Remove bundled OpenSSL
Having a bundled version of OpenSSL (header files and precompiled
binaries) is not a good thing. While it can be convenient in specific
cases, it has a number of drawbacks (possible compatibility issues with
different runtime environments -- Glibc vs. musl etc.; does not scale as
new architectures are added ;  takes significant space in the repository
(compared to source code anyways).

Therefore remove the OpenSSL headers and libraries and expect the user
to provide a valid OPTEE_OPENSSL_EXPORT path when WITH_OPENSSL is y.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>